+++ README.md
@@ -1,2 +1,24 @@
 # Vanilla JS Toolkit
 
+## TypeScript
+
+- CLI - https://www.typescriptlang.org/docs/handbook/compiler-options.html#handbook-content
+- TypeScript ESLint - https://typescript-eslint.io/docs/
+
+## Install
+
+```bash
+npm install
+```
+
+## Build
+
+```bash
+npm run build
+```
+
+## Example
+
+```bash
+node dist/examples/index.js
+```
