+++ package.json
@@ -7,7 +7,8 @@
     "example": "examples"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "rm -rf ./dist && ./node_modules/.bin/tsc",
+    "lint": "./node_modules/.bin/eslint ./src"
   },
   "repository": {
     "type": "git",
@@ -18,5 +19,11 @@
   "bugs": {
     "url": "https://github.com/WanuziaLab/vanilla-js-toolkit/issues"
   },
-  "homepage": "https://github.com/WanuziaLab/vanilla-js-toolkit#readme"
+  "homepage": "https://github.com/WanuziaLab/vanilla-js-toolkit#readme",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "5.36.2",
+    "@typescript-eslint/parser": "5.36.2",
+    "eslint": "8.23.0",
+    "typescript": "4.8.2"
+  }
 }
