+++ src/helpers/array.ts
@@ -1,0 +1,21 @@
+import { isArray } from './is'
+
+interface IData {
+  length: number
+  slice: (start: number, end: number) => unknown[]
+  data?: unknown[]
+}
+
+type ICallback = (value: IData) => void
+
+export const splitArrayIntoChunks = (data: IData, parts = 1, callback: ICallback) => {
+  if (!isArray(data)) return data
+
+  const length = data.length
+  const chunkSize = Math.ceil(length / parts)
+
+  for (let index = 0; index < length; index += chunkSize) {
+    const chunk = data.slice(index, index + chunkSize)
+    callback && callback(chunk)
+  }
+}
